add toasts and modal

+++ src/components/dashboard/all-teachers.tsx
@@ -1,17 +1,27 @@
-"use client"
+"use client";
 
 import React from "react";
 import { Button } from "../ui/button";
 import { useRefreshStore } from "@/store";
-import Link from "next/link";
-import { redirect } from "next/navigation";
 import ScrollableFeed from "react-scrollable-feed";
+import { useToast } from "@/components/ui/use-toast";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 
 const AllTeachers = ({ teachers }: { teachers: TeacherType[] }) => {
   const refresh = useRefreshStore((state) => state.refresh);
   const setRefresh = useRefreshStore((state) => state.setRefresh);
 
-  
+  const { toast } = useToast();
 
   const deleteTeacher = async (id: number) => {
     try {
@@ -23,21 +33,28 @@ const AllTeachers = ({ teachers }: { teachers: TeacherType[] }) => {
         body: JSON.stringify({ id }),
       });
       const data = await response.json();
+      toast({
+        title: data.msg,
+        variant: "success",
+      });
       setRefresh(!refresh);
-      console.log(data);
     } catch (error) {
       console.error(error);
+      toast({
+        title: "Could not delete teacher",
+        variant: "destructive",
+      });
     }
   };
 
-  const handleTable = (id : string) => {
+  const handleTable = (id: string) => {
     if (id.length === 0) {
-      alert("No table assigned")
-  } else {
+      alert("No table assigned");
+    } else {
       const url = `/table/${id}`;
-      alert(url)
-  }
-  }
+      alert(url);
+    }
+  };
 
   return (
     <div className="space-y-5">
@@ -46,29 +63,56 @@ const AllTeachers = ({ teachers }: { teachers: TeacherType[] }) => {
           <h3 className="scroll-m-20 text-2xl font-semibold tracking-tight">
             All Teachers
           </h3>
-            <ScrollableFeed className="h-96">
+          {/* <ScrollableFeed className="h-96"> */}
           <div className="flex flex-wrap sm:gap-24 gap-5 justify-around w-full">
             {teachers.map((teacher) => (
-              <div key={teacher.id} className="space-y-5 grid gird-cols-3 p-5 rounded-lg sm:w-1/4 w-full bg-slate-100">
-                <span>{teacher.firstName} {teacher.lastName}</span>
+              <div
+                key={teacher.id}
+                className="space-y-5 grid gird-cols-3 p-5 rounded-lg sm:w-1/4 w-full bg-slate-100"
+              >
+                <span>
+                  {teacher.firstName} {teacher.lastName}
+                </span>
                 <div className="flex justify-between">
-                <Button
-                  onClick={() => {
-                    deleteTeacher(teacher.id);
-                  }}
+                  <AlertDialog>
+                    <AlertDialogTrigger>
+                      <Button>Delete</Button>
+                    </AlertDialogTrigger>
+                    <AlertDialogContent>
+                      <AlertDialogHeader>
+                        <AlertDialogTitle>Are you sure?</AlertDialogTitle>
+                        <AlertDialogDescription>
+                          This action cannot be undone. This will permanently
+                          delete teacher and remove data from our servers.
+                        </AlertDialogDescription>
+                      </AlertDialogHeader>
+                      <AlertDialogFooter>
+                        <AlertDialogCancel>Cancel</AlertDialogCancel>
+                        <AlertDialogAction
+                          onClick={() => {
+                            deleteTeacher(teacher.id);
+                          }}
+                        >
+                          Continue
+                        </AlertDialogAction>
+                      </AlertDialogFooter>
+                    </AlertDialogContent>
+                  </AlertDialog>
+                  <Button
+                    variant={"primary"}
+                    onClick={() => {
+                      handleTable(teacher.table);
+                    }}
                   >
-                  Delete
-                </Button>
-                <Button variant={"primary"} onClick={()=>{handleTable(teacher.table)}}>
-                  Table
-                </Button>
-                  </div>
+                    Table
+                  </Button>
+                </div>
               </div>
             ))}
           </div>
-            </ScrollableFeed>
+          {/* </ScrollableFeed> */}
         </div>
-      ) }
+      )}
     </div>
   );
 };
